More fixes for signals related to changing the View.
The "view changed" signal had too many roles.  Added a new "completed
view change" signal that is issued only after a change is complete
and should be logged.  "view changed" is issued in the middle of a
mouse move, but "completed view change" is not.

Also deleted some old code.