Changed the relationship of clipping plane normals to vtk's InsideOut flag.
The old way was giving unintuitive results when multiple clipping
planes were used.  Points were displayed if they were on the correct
side of *any* clipping plane.  Now they have to be on the correct
side of *all* clipping planes.

Added a boolean constructor arg to the PlaneAndArrowLayer, which
controls the sign of the arrow.  It had to be flipped when the
InsideOut change was made, but the widget shouldn't have to know
about what it's being used for.  Now it can be used in hypothetical
contexts in which the arrow should be displayed on the other side
of the plane.