Progress on adding higher order elements.
Most of what CSkeletonElement::realElement used to do is now done
by MasterElement::build, which is a more natural place for it.
build now constructs the Nodes for all of the ProtoNodes, not just
the corners.

Nodes on edges and faces that are shared by elements are stored in
SkelElNodeMap structures that are passed down to MasterElement::build
from CSkeleton::populate_femesh.  This allows communication between
elements.

Got rid of (commented out) ElementExteriorNodeIterator.  It was
only used by some 2D-only display methods and wasn't compatible
with the current changes to Element and MasterElement.
Removed Element::nexteriorfuncnodes.

Removed a lot of 2D-only code from masterelement.*.