Fixed categoryVolumes, which shouldn't throw an exception if the volumes don't add up.
Copying a fix from OOF2, more or less, for a bug that hasn't arisen
yet in OOF3D but would have.  If a node is illegally moved outside
the Microstructure, then it may create geometrically legal elements
that are partially outside, and the volumes of the voxel categories
in those elements won't add up to the total volume of the element.
This should *not* raise an exception.

I deleted the exception and put the whole volume check inside
an #ifdef DEBUG block.