Added more machinery for higher order elements.
Subclasses of MasterElement3D have two new methods,
getIntermediateEdgeNodes and getIntermediateFaceNodes, which are
used to find the correct set of nodes to use when creating adjacent
edge and face elements.

When Elements are created, they have access to a map of nodes created
on other edges and faces, so that they can find shared nodes.

Added a typedef NodeIndexVec used to hold indices of nodes and
defined it in masterelement_i.h, which may or may not be the right
place for it.

This version compiles and passes solver_test.py, using linear
elements.  It hasn't actuall been tested on higher order elements
yet.