Progress on higher order elements.  TET4_10 compiles and runs, but gives wrong answers.
Changed node order in TET4_10 to match VTK_QUADRATIC_TETRA so that
we can plot data without switching node order.

FEMesh::getGrid doesn't assume that the elements are linear and
gets the vtk cell type and number of nodes from the Elements.

Added 3rd order gauss points for the tetrahedron.

Added yet another static table of topological data, edgeNodes, to
CSkeletonElement.  It gives the nodes at either end of an edge.

Added operator== for heterogeneous IndexTypes.  It had operator!=
but not operator==.