Fixed a typemap that wasn't checking for errors properly, and added a way of looping over values of a Python Enum.
The SpaceIndex "in" typemap wasn't checking that its argument was
an Enum class, and a new test in mesh_test.py was passing an int
instead of an Enum.  The error was difficult to find because
subsequent Python API calls were resetting the error state.

Now Enum classes have a .values member, so they can be looped
over (for x in enumclass.values: ....) and mesh_test.py uses
it.