Added tests for interpolation, integration, and differentiation on 2D elements, and fixed some bugs.
The Element::Jdmasterdx functions were assuming that the dimension
of space was the same as the dimension of the element.  So was
ShapeFunction::realderiv.

Swigged masterCoord2D() and masterCoord1D() so that they can
be used in tests.

Fixed an error in the gauss points for triangular elements.

Added a bunch of tests to mesh_test.py for the 2D elements, and
integrating higher order polynomials on 3D elements.