Added type checking to the AutoNumberWidget.
Also added subclasses AutoFloatWidget and AutoIntWidget.

This fixes a bug that was making the dialog box for new
ScheduledOutputs spew a lot of errors when the check box for
an AutoIntParameter was clicked and the widget's text field
was empty.

I'm not sure why this isn't a problem in OOF2.  Maybe it's because
in OOF2 the various ingredients for a ScheduledOutput are set in
different dialogs, so the self consistency check doesn't require
all of them to be readable at once.