Fixed a bug in AutoGroup, and added a comment to the tutorial where it's used.
When discussing the Activity Viewer, the basic tutorial says to run
Autogroup on an image with thousands of different colors, which
takes a long time and uses a lot of memory.  If the user doesn't
stop the procedure, it may run out of memory.  The tutorial mentions
this now.

The autogroup() function wasn't checking progress.stopped() frequently
enough, and would appear to hang.