Modifications to "Fix Illegal Elements".
Fix Illegal was trying to move a node to the average position of
its neighbors, even if the node was on a face and couldn't legally
move to the average position.   Now it computes the average using
only the neighbors that are on the same plane(s), if the node is
constrained.  This allows it to fix some elements that were
unfixable before.

Also, the time stamps for the checks for suspect and illegal elements
were being incremented incorrectly. Because the way the virtual
functions were set up, the time stamps were incremented in the
sheriff skeleton.  Now the functions all are non-virtual in
CSkeletonBase, which uses its virtual methods for looping over
elements.