Fixed(?) a voxel group bug.
Groups weren't being sorted correctly in some circumstances so
sometimes removing voxels failed.  The (?) is because I don't
understand why this version works and the previous one didn't,
but I'm committing this anyway and will try interpolating
between the versions to figure out what's going on.

Changed operator<< for std::vectors so that it inserts line
breaks in long vectors.  Printing the contents of large pixel
groups was confusing the terminal.