Divide the Microstructure into subregions and create a VSBGraph for each.
Dividing the Microstructure into subregions allows the expensive
clipping operation on VSBGraphs to be skipped for large parts of
the Microstructure when computing Element homogeneities.  The
computation is only done where the bounding box of the element
intersects the subregion.

Currently the size of the subregion is hard coded, and the bounding
boxes of all subregions are inspected.  Neither of those is optimal.
However, the code is already about a zillion times faster than it
was before for large Microstructures.