Fixed problems in splitting voxel set boundaries into subregions and in clipping them to compute element homogeneity.
The r3d graph clipping algorithm has some peculiarities if the graph
is not 3-vertex connected.  It's easy to create such a graph from
a polyhedron of voxels (eg, a 3x3x3 cube with one voxel removed
from the center of one edge).  The clipped graph can contain pairs
of nodes that create a two sided (degenerate) loop.  Such a loop
can and should be ignored, by connecting together the other two
nodes that are neighbors of the nodes in the loop.   Now
VSGraph::clipInPlace and VSBGraph::copyAndClip run removeDegenerateFaces
after connectClippedNodes.

The size of the subregions can now be set by menu commands.  That's
presumably temporary, since it should be automatic.  A negative or
zero size means that subregions won't be used.

There's lots of commented out debugging code in here which will
have to be cleaned up.