The VoxelSetBoundary graphs are now divided into bins of approximately the average element size.
Dividing the VSBs this way makes them more efficient, because only
the bins that overlap an element's bounding box need to be examined.
Since the size depends on the skeleton element size, the VSBs are
now stored in the Skeleton, not the Microstructure, and a bunch of
methods had to move there with them.  This means that lots of methods
that used to take a CMicrostructure* argument now take a CSkeletonBase*
instead, which accounts for most of the files in this commit.

Removed the temporary menu commands for setting the VSB bin size.

Removed the checkTopology feature from VSB methods, since it was
unreliable for concave polyhedra.

Cleaned out some old debugging code.