Fixed a problem with timestamps that was causing homogeneity to be recomputed unnecessarily, and added a C++ debug() method.
A bunch of timestamps related to homogeneity calculation and node
motion weren't being copied when Skeletons were copied, so every
operation that starts with a copied Skeleton was recomputing the
homogeneity of every element.

Added a TODO about making CSkeletonMultiNodeSelectable::nodes not
be a pointer anymore.

The new debug() method in C++ now returns the same value as
debug.debug() in Python.   Previously we could turn debugging
on and off in C++ only at compile time.  Now code blocks can be toggled
any time via the Debug menu if they're inside an "if(debug()) {...}"
block.