Modified the "Tumble" algorithm so that objects are rotated about the center of the visible points.
Mouse down in tumble mode computes the center of rotation.

Canvas layers that care about the center of rotation should redefine
OOFCanvasLayerBase::visibleBoundingBox() and compute the bounding box
of their visible region.  They may use the utility function
getVisibleBoundingBox() in canvaslayers.C to do so if their data is
in the form of a vtkDataSet.

Made OOFCanvasLayer::showing() virtual, and moved it to the base class.
visibleBoundingBox needed it for some overlayers.

Changed the null constructor for CRectangularPrism_ so that it sets
the max values to -std::numeric_limits::max() and the mins to
+std::numeric_limits::max().  Then if a bounding box is being computed
by repeated calls to CRectangularPrism_::swallow(), no special treatment
is required for the first point.

Swigged CRectangularPrism::center and ::swallowPrism, and changed
the swigged constructor to the null constructor.

Added a SmallMatrix multiplication operator for Coord3Ds. The
center of rotation code would have to translate between Coord3Ds
and DoubleVecs otherwise.

Added a rotateAboutAxis() routine to smallmatrix.C that generates
a 3x3 rotation matrix given an axis and an angle.