Improved the new Tumble algorithm.
The elevation and azimuthal rotations are now applied separately,
and the view-up vector is recomputed before applying the second
rotatation.  Applying the rotations together was making the
rotated object walk around the screen if the angles were large.

Added visibleBoundingBox methods to more OOFCanvasLayer subclasses
so that they're rotated correctly.

Added some TODOs and comments.