Miscellaneous minor changes.
Turned off the TESTCOVERAGE macro in crefine.[Ch].

Changed the sorting order of the graphics layers for the
clipping plane and voxel selection widgets, and made their
names consistent.

Some places in which a Coord3D was being passed to a vtk function
that expected a double* were using Coord3D::xpointer to convert,
but Coord3D has a conversion operator so the explicit call wasn't
necessary.  I didn't try to change this everywhere, though.

Added some debugging lines to figure out why tumbling the image
is so slow.  Recalculating the center of rotation is very slow
for large images.  I'm not sure what to do about it.

Added some TODOs.