Modifer keys on tumble mouse clicks control what the center of rotation is and when it's recomputed.
Computing the center point of the visible objects can be slow, so
instead of figuring out how to make it faster, for now I've just
made the calculation optional.  The center is only recomputed if
the shift key is held down.  If the control key is held down, the
focal point is used.  If neither key is pressed, the last center
of rotation is used.