Some switchboard callbacks weren't being cleared when a graphics window was closed.
The SkeletonSelectionToolboxGUI wasn't relaying the closed call()
to its enclosed SkeletonSelectionToolboxModeGUIs.  Fixing this
didn't fix the bug I'm trying to track down, which is that the
program crashes when a new graphics window is opened after another
one has been closed.

Also added a Dump Switchboard command in the Debug menu, which prints
all of the requested switchboard callbacks.