Fixed a bug triggered by clicking on nothing in the canvas.
Some of the findClicked... methods in GhostGfxWIndow were returning
a tuple when successful and just None when they failed.  Now they
return an appropriately sized tuple of Nones, so that the calling
methods don't have to check the result before unpacking it.  The bug
was actually that not all of the callers were checking, but now none
of them do.

Also started debugging a problem where vtk is printing unnecessary
warning messages when a data set is empty.