This version works with vtk 7.1.1 and quartz/cocoa (not X11!) on Mac.
This version hasn't yet been tested on X11, so it may need some
tweaks.

Added OOF_USE_COCOA as a preprocessor definition in C++ files.

oofcanvas3d.C inserts a vtkCocoaRenderWindow into a gtk drawing area.

The new file oofcanvas3d.mm just #includes all of oofcanvas3d.C.
.mm is the suffix for ObjectiveC files that also contain C++.  When
building the quartz/cocoa version, the gdk header files lead to
Apple ObjectiveC files.  Fortunately, compiling oofcanvas3d.C
with clang as a .mm file seems to be sufficient.
common/IO/GUI/DIR.py includes oofcanvas3d.mm instead of oofcanvas3d.C
when --cocoa is provided as an option to setup.py.

Added a hack in setup.py so that distutils will recognize .mm as
an ObjectiveC suffix.

Including the quartz/cocoa header files introduced a name conflict
for ThreadState and ThreadID.  I changed them to OOFThreadState
and OOFThreadID.

Changed the Questioner dialog so that it's not a subwindow of the
main OOF GUI window, which was causing it to appear behind another
window if the main window wasn't in front.

Removed the pre-python 2.5 compatibility code that setup.py was
inserting into oofconfig.h.  Made setup.py a bit more verbose about
what it's doing with oofconfig.h.

oof2-build now always removes the oof3d script from the build directory
and rebuilds it.  Not doing that meant that oof3d would run with the
old version of python if a new version of python was installed, because
the version used to build oof3d was stored in the #! line of oof3d,
and the file wasn't being replaced.  (This mattered when I was switching
between x11 and cocoa macports installations.)