Fixed a problem with loading images from a data file.
When we use vtkImportImage to construct an OOFImage3D from an array
of chars, we have to make sure to keep a copy of the importer, or
the char array is lost.  This may be a difference between vtk5 and
vtk7.  Now there's a subclass of OOFImage3D that keeps the importer.