Fixed a bunch of vtk issues.
Fixed oofOverlayVoxels so that it works on images with alpha
channels again.

Fixed FEMesh::getMaterialCellData, which was returning the wrong
vtk object.  Perhaps vtk5 was willing to convert a vtkIntArray smart
pointer into a vtkDataArray smart pointer implicitly.

Fixed MaterialImage, which needed to explicitly update its vtk pipeline.

Changed the parameters of the Blur image modifier to
PositiveFloatParameters, and set the default value of 'radius' to
1, instead of 0.  It may have behaved differently in vtk5, or,
more likely, we didn't notice that it didn't work with 0.