Added some debugging code.
This branch is to figure out why opening two graphics windows and
closing the first one with the window manager close button crashes
the program.  It only happens on Linux.  It doesn't happen if the
second one is closed first, or if the "File/Close" menu item is
used, or if ^W is typed.