Still debugging the crash that occurs when closing the graphics window on Linux.
The previous version crashed in two ways, depending on how the window
was closed.  A debugging version of GhostOOFCanvas::deactivate() had
been deleting the vtkRenderer too soon.

It now seems clear that the problem is with the OOFCanvas3D destructor,
or at least with the order of destruction of the canvas and the vtk
components.