Still fiddling with graphics window shutdown code.  It works with vtk8 on Ubuntu.
vtkXOpenGLRenderWindow::Finalize isn't called explicitly within the OOF
code any more.  The render window destructor calls it.

The OOFCanvas3D destructor now sets the render window's WindowId to 0.
I don't know if this is correct, but it prevents the errors that otherwise
occur within the render window's destructor.

This version appears to work with vtk8 on Ubuntu.  I haven't tested it
with vtk7 yet.

Cleaned up some signal handling code in oofcanvas3d.C.  The mouse event
signal is handled like the other signals.   All of the callbacks return
FALSE so that events propagate to the parent callbacks.  I don't know
if that's important in this context.