Added a call to XSync when closing gfx windows.
This seems to sort of fix an intermittent crash that occurs when
closing graphics windows on Linux.

When using vtk8, the X11 error message from the crash says that
probably XInitThreads wasn't called, but adding XInitThreads
doesn't help. I left in the code that calls it, but it's inside
an #ifdef/#endif block that ensures that it doesn't get called,
with comments explaining why.