The contents of the graphics window don't disappear after resizing on Linux.
For reasons that I don't understand, the contents of the graphics
window would disappear after resizing the window on Linux.  Calling
vtkXOpenGLRenderWindow::Render would make them reappear, but not
if it was called from within the 'configure' or 'expose' event callbacks.
Now the expose callback installs an idle callback that calls Render,
and the window contents don't disappear.