Removed calls to vtkRenderWindow::SetAAFrames and replaced them with SetMultiSamples.
SetAAFrames is deprecated.  SetMultiSamples seems to work, but can't
be reset unless the canvas is rebuilt.  I'm committing this now
so that I can see if it behaves the same way on Linux, since it's
not documented.

The Antialias menu item is now Multisampling, and takes an int which
is the arg to the vtk SetMultiSamples method, whatever that might mean.
I may give it a reasonable fixed value and just expose a bool to the user.