Changed the Antialias code to either vtk's FXAA or SSAA, depending on a compiler flag.
I asked the vtk users list about the correct substitute for the
deprecated vtkRenderWindow::SetAAFrames method.  There are two,
other than multisampling, which I already tried, and which seems
to rely on hardware.  FXAA is cheap, and SSAA is less so but
supposedly still faster than AAFrames.

FXAA seems to work, but the results aren't as pretty as AAFrames,
but I haven't played with its parameters yet.  SSAA is doing very
strange things to the surface representation of grids.

To use SSAA, #define OOF_USE_SSAA_ANTIALIASING.

The MultiSampling antialiasing is also still available.