Antialiasing is working with the vtk SSAA method.
Removed calls to the vtk MultiSampling antialiasing method, which
didn't work unless the canvas was rebuilt, which is a pain..

Added methods for using vtk FXAA antialiasing, which looked much
worse than SSAA, so they are commented out.

SSAA works properly on OS X only using the very latest vtk 8.1.x.
At least, it doesn't work in 8.1.0.rc2, but I reported the bug and
it's been fixed in the master.  I don't know if the fix will be in
the final 8.1.0.  I also haven't tested it in 8.0.1, 7.1.1 or
earlier.

Added a comment about using vtkNew instead of vtkSmartPointer
in ghostoofcanvas.h.

Removed an unused GhostOOFCanvas::rendered flag.