Trying to fix situations in which the vtk render window is misaligned with the gtk drawing area.
Added separate methods that reposition and resize the render window
but it's not clear where to call them.

The window get misaligned when a View is changed via a scripted command,
probably because the View contains size information and resizes
the render window.  This was done to make mouse clicks reproducible.
I will experiment with removing the resizing...

Removed some obsolete code and fixed some comments.