Fixed a couple of bugs related to resizing the graphics window.
The canvas size needs to be contained in the View objects
so that mouse clicks can be interpreted correctly.  But if
a script contains a command that changes the View (tumbling, panning,
etc) it *shouldn't* change the canvas size, since the script
might be executed with a differently sized window.  Now
GhostOOFCanvas::set_view_nolock and GhostOOFCanvas::restore_view
have a boolean argument that says whether or not they should
respect the size in their View argument.  If they're called
by GhostOOFCanvas::set_view they don't use the size, but if they're
called by one of the many internal click-interpretation routines
they do.

Removed GhostOOFCanvas::matchSize(), since it's unnecessary.