Some skeleton and mesh displays weren't updating when their filter's output changed.
The mechanism that contour displays had been using to trigger an
update when a filter's output changed is now in the SkelMeshDisplayMethod
base class.  SkeletonFilterRegistrations (also used for Meshes) can
set switchboard callback functions.  If a filter relies on external
objects that may cahnge (such as selections or groups) it can get
a notification of the change and can send "skeleton filter changed"
which is received the SkelMeshDisplayMethod.