Added a menu item for toggling the factor of two needed to workaround the gtk/vtk window scaling bug.
The bug only seems to occur on Mac laptops and presumably has something to
do with the graphics drivers.

Also starting working on machinery for raising and lowering layers,
which will rely on vtk's coincident geometry resolution machinery.
Either I don't understand it or it's buggy.

As part of that, added a way for canvas layers to report how many
vtkActors they use, so that they can all be raised or lowered
together.  This may not be necessary.