Layer ordering and reordering is now implemented using the vtk coincident topology machinery.
GhostGfxWindow.setLayerOffsetParameters is called after anything
changes the layer ordering.  It calls each CanvasLayer's new
setCoincidentToplogyParams method.

Added a new LayerOrdering object, Celestial, to be the partner of
Abysmal.  Abysmal is always on the bottom, Celestial is always on
top.

There's something wrong with drag and drop in the layer list, and
also with mainaining the proper selection when the list is rearranged.