Fixed a bug that was making the viewer toolbox incorrectly list some edited clipping planes.
If a plane was edited and only the class of the normal was changed
(eg changed from CUnitVectorDirection to CAngleDirection) but the
actual mathematical direction wasn't changed, then the list of
clipping planes in the toolbox wouldn't update, because it thought
that nothing had changed.  Added a CDirection::identical() method,
which is like operator== but also compares the class names, and
used it in ClippingPlane::operator==().