Working on the 3D widgets some more.
Changed the name of BoxAndArrowLayer for selecting rectangular
regions to BoxWidgetLayer and deleted all of the arrow code.  The
user can click and drag on the sides of the box just as well, and
won't be re-orienting them.

Changed the name of ClipPlaneClickAndDragDisplay to ClippingPlaneWidget,
which is shorter and more informative.  The name appears in the gfx
window's layer list.