Working on bugs in the widget for selecting rectangular regions of voxels.
Committing now in a sort of half baked state because of the impending
government shutdown.

The RectangularPrismSelectorGUI isn't working.  It now hides and/or
dims correctly when the toolbox isn't active, but vtkCellPicker::Pick
doesn't pick anything.  I thought it had something to do with hiding
and showing the layer when I was tracking down the problem by copying
differences one by one between the previous commit on this branch
and the development branch.  But now I'm not so sure.  Hiding and
showing is now robust and is just like the ClippingPlaneWidget (I
think) but that one works and this one doesn't.

All gfx toolboxes now get their name from their associated non-gfx
toolbox.  There had been some inconsistencies, and this was causing
confusion in some switchboard calls that were using toolbox names.

Minor code cleanup in RectangularPrismSelectorGUI.  Also it now
doesn't accept move events if the mouse isn't down.

Removed the set_visibility methods from PlaneAndArrowLayer
and BoxWidgetLayer.  They should have been using the generic
show() and hide() in the base class, or the new DisplayMethod.setVisibility().

Added DisplayMethod.setVisibility(bool) for situations in which
it is neater than if xxx: hide(); else: show().

Added setEmtpy(False) calls to the BoxWidget and ClippingPlaneWidget's
newLayer methods.