Fixed a bug in the voxel region selection widget (BoxWidgetLayer).
vtkPoints::Modified has to be called explicitly after editing
the vtkPoints in a grid.  Various BoxWidgetLayer methods that
move the points weren't calling it, so the vtkPicker was using
the bounding box of the unmodified points.

Added calls to vtkPoints::Modified in other spots where points
are moved.  It might not be necessary in all cases, but it can't
hurt.

Deleted debugging lines.