Working on the voxel selection interface.
Instead of starting the rectangular prism selection by clicking on
the canvas, there's now a "Start" button in the toolbox.  This is
more intuitive.   There's also a "Cancel" button.  The "Done" button
is still not hooked up to anything.

The VoxelRegionSelectionDisplay can now be hidden or dimmed when
it's inactive.

The classes for selecting voxels need to be refactored.  The
distinction between VoxelRegionSelectWidget and RectangularPrismSelectorGUI
is unnatural.  They should be merged into one class, and then the
generic bits (like the mouse handling mechanism) that will be useful
in other selection methods should be separated out or moved into
PixelSelectToolboxGUI.