More fixes to abaqus output, skeleton homogeneity, and mesh creation during homogeneity calculation.
newMesh in meshmenu.py acquires the skeleton read lock, so that a
mesh won't be created while homogeneiety is still being computed.

Formatting fixes to Mesh abaqus output.

Minor changes to Skeleton abaqus output.  It's not terribly slow,
unlike the Mesh output was, so I'm not moving the code to C++.

Backported a fix from 3.l.later that prevented categories from being
recomputed far too often, which was making it very slow to compute
homogeneities and create skeletons.