Added a --makedepend option to setup.py, which no longer computes dependencies automatically.
Dependency calculation is slow, so now it's optional.  If you change
which files are #included in other files, or add new C++ source files,
run setup.py or oof2-build with the --makedepend option.  Otherwise
the dependencies are assumed to be the same as they were the last
time the program was built, and are read from a cached file.