Progress on the new generic selection toolbox machinery.
The GUI for selecting a rectangular prism of voxels is nearly complete.
Voxels aren't actually selected yet, though.

common/IO/GUI/mousehandler.py is obsolete and is being replaced by
common/IO/mousehandler.py.  The MouseHandler subclasses keep track
of mouse buttons and click positions, but expect to pass their
results on to a more concrete object (such as VoxelRegionSelectionDisplay)
that will do something useful with the data.

BoxWidgetLayer::set_box takes either a Coord3D or a CRectangularPrism
argument, and only the CRectangularPrism version is swigged. The
Coord3D version assumes that one corner is at the origin.  This
change allows the widget to be initialzed using its previous state,
which might not have had one corner at the origin.

VoxelRegionSelectionDisplay communicates directly with its controlling
toolbox, instead of using switchboard signals.  It keeps more state
data so that it can be restored if mouse handling is temporarily passed
to a different controller.

Added a Reset button to the generic selection toolbox, and a reset
method to VoxelRegionSelectionDisplay.

Added a CRectangularPrism constructor that takes two Coord* args,
so it can be called from python.

VoxelSelectionOp subclasses can specify a widgetType, which is the
class used to create the GUI widget for editing the selection in
the canvas, and an "instructions" string which is displayed in the
toolbox.