Starting a new branch for developing simpler generic selection toolbox machinery.
"Simpler" means simpler than the previous attempt in the
new-generic-selection branch.

This attempt uses a RegisteredClassFactory for the selection method
chooser, unlike the previous one.  It copies the new MouseHandler
class from the previous iteration, so that subclasses of
SelectionMethodGUI don't have to implement it themselves.

Every RegisteredClass subclass now contains a reference to its
Registration.

Assigning SelectionMethodGUI subclasses to their SelectionMethod
subclasses is now done with a decorator defined in
pixelselectionmethodGUI.py (although it will be moved out of there
eventually because it can be used elsewhere). The decorator adds a
'gui' member to the subclass's Registration, so the gui class can
be found easily.