The new generic selection toolbox machinery seems to be working, although...
... it's only been implemented for selecting rectangular prisms of voxels
so far, and the actual selection isn't done.  But the correct widgets are
brought up in the toolbox's RegisteredClassFactory and in the canvas,
and they behave correctly when switching toolboxes or closing the window.

Selection methods should have a SelectionMethodGUI class, which is
associated with the SelectionMethod by the selectionGUIfor decorator.
The SelectionMethodGUI is in charge of creating gtk widgets for the
toolbox and connecting to a canvaslayer and coordinating the two.  It also
interprets the mouse events.