Progress on the new selection toolbox architecture.
Removed the Historian from the generic selection toolbox.  What it
means to repeat an operation is badly defined, since the view might
have been rotated since the operation was recorded.

Clicks can now be resolved on multiple canvas layers, so it's not
necessary to fully specify which Who object is being clicked.  Added
supporting functions GhostGfxWindow.allWhoClassLayers, which returns
all layers containing the given Who classes, and
GhostGfxWindow.findClickedCellCenterMulti, which is just like
findClickedCellCenter, but it finds the closest click point among
multiple layers.  Also GhostOOFCanvas::findClickedCellCenterMulti
and GhostOOFCanvas::findClickedCelllMulti_, which can be used to implement
multilayer versions of all of the earlier findClickedXXXXX methods.

SingleClickVoxelSelectionMethod.getVoxel now returns a tuple containing
the Who object that was clicked on and the position of the voxel.

The RCF in the selection toolbox now always lists all methods,
rather than trying to decide which methods could be applicable
to the "top" layer, since it's possible to click on layers that
aren't on top.

Changed getSource to getSelectionSource in the toolbox and gui toolbox.
getSelectionSource returns the Who object that contains the selection
object that's being modified.  That's not the same as the thing that
was clicked on to make the selection.