Removed the canvas overlayer machinery and added a separate vtk actor for displaying selected voxels.
Using overlayers to display selected voxels made sense in 2D, but
not so much in 3D.  It didn't work well if filters were used to
make parts of two bitmap objects visible.  Now that the vtk coincident
topology stuff is working, selected voxels can be displayed by
superimposing a layer of translucent cells, and we can let vtk sort
out the details.

Deleted the BitmapOverlayDisplayMethod, but for now have kept the
BitmapOverlay class.  It's still used to keep track of which voxels
are selected, but it is more complex than necessary.  Added a TODO
about removing the class and using something more appropriate in
CPixelSelection.