Moved the 'operator' param (Select, Toggle, etc) into the VoxelSelectionModifier and VoxelSelectionMethod classes.
'operator's value is an instance of a PixelSelectionOperator subclass.
Those subclasses are now defined in common/selectionoperators.py,
instead of pixelselectionmod.py, because they will eventually be
used for other kinds of selections.

It doesn't make sense for 'operator' to be set independently, because
it doesn't apply to all selection methods.

One problem with moving 'operator' into the RegisteredClass parameters
for the selection methods is that it appears in the selection
toolbox, even for single-click selection methods, where it's more
naturally set by keyboard modifier keys.  So I used the Parameter's
set_data method to hang a 'passive' string on the Parameter which
tells the ParameterTable to make the widget inactive.  This is a
bit sleazy because the individual widget, not the ParameterTable,
should be checking whether or not the widget is passive, but doing
that properly would entail modifying all of the Parameter classes.
That can be done later.

The passive widget is updated when the mouse is clicked.  It would be
good if it were updated when the modifier keys were pressed.  Doing
that entails getting keyboard events from the gtk drawing area (I think).
That's not working at the moment, so I'm committing now so that I can
check it on a system that's not using the quartz version of gtk.