Control and shift keys now change the passive selection operator widgets in the selection toolbox in real time, not just when the mouse is clicked.
Added a "modkeys" event to the mouse handler machinery, so toolbox
widgets can be notified when the currently pressed modifier keys
change.  Currently only control and shift are monitored but it would
be easy to add others.  Reporting non-modifier key presses would be
possible with a bit more work.

The selection operation widget for the Box selection method is now
misbehaving.  It's being set as if it's a passive widget sometimes.

Changed the names of the selection operators.  "Select Only" is now
"Select", and "Select" is "AddSelection".