Fixed a bug that was changing the 'operator' widget for the box selection method incorrectly.
The GenericSelectToolboxGUI was changing all parameter widgets in the
toolbox gui from its setParamValues method, when it was only
supposed to be setting the ones in its arg list.  The others were
being reset to the values in the Registration.   Added a new
setParamValues to RegisteredClassFactory which only changes
the given parameters, instead of rebuilding the whole factory
from the registration.