Fixed some voxel selection bugs.
SelectionMethodGUI classes now have install() and uninstall() methods
that are called when the toolbox switches from one method to another.
RectangularPrismSelectorGUI uses them to ensure that it's in the
right state when starting up.

The menu item for non-simple selection operations now calls
selection.start(), so the selection subclasses don't have to worry
about it.  The callback for simple selection operations does *not*
call start(), since it's not appropriate for Undo or Redo.

Fixed Despeckle and Elkcepsed, which hadn't been updated for the
new architecture. Also SelectMaterialPixels and ColorRange.

Minor code cleanup and improved comments.