Combined the separate toolboxes for element selection, node selection, etc back into one.
The new skeleton selection toolbox uses the same scheme that the
old ones did for displaying the toolboxes for the various selection
modes in a single pane and switching between them with radio buttons.
It's simpler, though, because more work is done by tne generic
selection toolbox.

The subclasses for both the gui and non-gui parts of the skeleton
selection toolboxes are now generated automatically by looping over
the SkeletonSelectionModes.  The singleton instances of each
SkeletonSelectionMode subclass are created by
skeletonselectiontoolbox.initialize().