Working on using couriers for Skeleton selection operations.
SkeletonSelectionMethods are broken at the moment.  Fixing that
will be next.

Some SkeletonSelectionModifiers work with couriers, though.

Selecting groups of Skeleton nodes, faces, etc, works.  The
different Select/Unselect/Toggle menu items are now just
a single item, with an 'operator' argument, as is done with
voxel selections.

Selecting nodes from the set of selected elements, faces, or
segments also works.

Changed common/selectionoperators.py so that it is no longer
restricted to voxel operations.  The different kinds of selections
must now have an intersectionCourier() method that returns the
appropriate type of courier to use to compute a set intersection
with the current selection.

The Select callback for SkeletonSelections now has a try/except
block around the actual selection call, and undoes selection.start()
(calling selection.unstart()) if the selection fails.  This means that
the couriers don't have to call selection.start() themselves, but can
still safely raise an exception if they can't be applied, without
leaving an incomplete selection on the redo/undo stack.

CGroupTracker::get_group raises an exception if the group can't be
found.

Added a missing include to cskeleton2.h.  I don't know why it only
started being a problem now.