Selecting internal boundary nodes and nodes of named boundaries now works using couriers.
Added a StupidCourier class, which transfers any iterable set of
CSkeletonSelectable*s from Python to C++.  It's stupid because the
whole point of the couriers is to avoid copying long lists of objects
from Python to C++.  The boundaries are stored as lists of objects
in python, so there's no way to avoid copying the lists.

Added a typedef for std::vector<CSkeletonSelectable*> and a
swig typemap that converts any iterable python container of
CSkeletonSelectables to it.  It's possible that we should rewrite
a bunch of typemaps to work on any kind of iterable python container
instead of having them only work on lists or tuples.