Various fixes and cosmetic improvements to the voxel and skeleton selection machinery.
The single-click voxel selection methods now receive a iPoint in
voxel coordinates, instead of a Point in physical coordinates.  The
point is computed correctly even if the voxels aren't 1x1x1.

Added an Invert button to the voxel selection page and made the
Invert and Clear registrations secret so that they don't appear in
the RCF on that page.  Added a comment about how they don't have
to be registered classes at all.

Fixed GhostOOFCanvas::findClickedCellCenterMulti(), which wasn't
actually finding cell centers.  This was causing incorrect pixels
to be selected sometimes.

Made the position parameter widgets in the voxel selection
toolbox passive, since the user shouldn't edit them directly.

Fixed the skeleton info toolboxes, which had an up() method that
was still using the old arguments.

Changed the names of the skeleton selection methods and the label
on the skeleton selection page so that skeleton components and
voxels are handled similarly.  Eg, modifier names don't have
underscores, don't begin with "Select".

The order of selection modifiers is now consistent for
elements, faces, nodes, and segments.